+++ src/main/java/com/example/humorie/account/service/AccountService.java
@@ -130,8 +130,16 @@ public class AccountService {
         if (accountDetail.isPresent()) {
             return accountDetail.get().getAccountName();
         } else {
-            throw new RuntimeException("Account not found with email: " + email);
+            throw new ErrorException(ErrorCode.NONE_EXIST_USER);
         }
+    }
+
+    public String isAccountNameAvailable(String accountName) {
+        if(accountRepository.existsByAccountName(accountName)) {
+            throw new ErrorException(ErrorCode.ID_EXISTS);
+        }
+
+        return "The account name is available";
     }
 
 }
